refactor: rename date to PypDate component
