+++ components/CategoriesList/index.ts
