+++ pages/[city]/[category]/[number]/index.tsx
@@ -4,12 +4,12 @@ import Vidverto from "components/ads/vidverto";
 import { Aside } from "components/Aside";
 import Breadcrumbs from "components/breadcrumbs";
 import CTA from "components/call-to-action";
-import PypDate from "components/date";
 import { Header } from "components/Header";
 import Hours from "components/hours";
 import LicensePlate from "components/license-plate";
 import NumberLinks from "components/number-links";
 import Post from "components/post";
+import PypDate from "components/PypDate";
 import { format } from "date-fns";
 import markdownToHtml from "lib/markdownToHtml";
 import getPostBySlugs from "lib/posts";
