+++ pages/[city]/[category]/index.tsx
@@ -2,10 +2,10 @@ import cities from "@mauriciorobayo/pyptron";
 import TheMoneytizer from "components/ads/the-moneytizer";
 import { Aside } from "components/Aside";
 import CTA from "components/call-to-action";
-import { isValidDateString } from "components/date/utils";
 import DaysList from "components/days-list";
 import { Header } from "components/Header";
 import Post from "components/post";
+import { isValidDateString } from "components/PypDate/utils";
 import markdownToHtml from "lib/markdownToHtml";
 import getPostBySlugs from "lib/posts";
 import { AMERICA_BOGOTA, CityType, dateParts, isCity } from "lib/utils";
