+++ pages/[city]/index.tsx
@@ -3,8 +3,8 @@ import TheMoneytizer from "components/ads/the-moneytizer";
 import { Aside } from "components/Aside";
 import CTA from "components/call-to-action";
 import { CategoriesList } from "components/CategoriesList";
-import { getLocalLongDateString } from "components/date/utils";
 import { Header } from "components/Header";
+import { getLocalLongDateString } from "components/PypDate/utils";
 import { AMERICA_BOGOTA, CityType, dateParts, isCity } from "lib/utils";
 import { GetStaticPaths, GetStaticProps } from "next";
 import { NextSeo } from "next-seo";
